fix:  파싱 하거나 검증할때 Long 으로 바꾸어 잠재적인 오류제거

+++ src/main/java/calculator/numberExtractor/CustomNumberExtractor.java
@@ -31,7 +31,7 @@ public class CustomNumberExtractor implements NumberExtractor {
     private void saveValidatedNumber(String[] splitDelimiter, NumberRepository numberRepository) {
         for (String split : splitDelimiter) {
 
-            int parseNumber = parseNumber(split);
+            long parseNumber = parseNumber(split);
             validator.validate(parseNumber);
             numberRepository.saveNumber(parseNumber);
         }
